linalg.h compatability
A number of changes were needed here.  First pinv_array is the only
active function in the interface file.  Everything elese is
called/generated from this.  A 1d array is used for Ax in pinv_array so
the calling function (it only looks like one) should flatten (rowwise)
with ravel as the docstring indicates.

std::complex is now used instead of npy_cfloat_wrapper.  c++98 compliant
instead of c++11.